031724_update

+++ Inference_pytorch/NeuroSIM/InputParameter.h
@@ -44,7 +44,8 @@
 class InputParameter {
 public:
 	/* Properties */
-	int processNode;				/* Process node (nm) */
+	// CFET
+	double processNode;				/* Process node (nm) */
 	DeviceRoadmap deviceRoadmap;	/* ITRS roadmap: HP or LSTP */
 	int temperature;				/* The ambient temperature, Unit: K */
 	TransistorType transistorType;	/* Conventional CMOS, 2D FET, or TFET */
